AoC 2023/11

+++ rust/src/advent_of_code_2023/advent_2023_11.rs
@@ -1,7 +1,7 @@
 use crate::utils::read_lines;
 use itertools::Itertools;
 
-fn solve(factor: i64) -> usize {
+fn solve(factor: usize) -> usize {
     let l = read_lines("advent_2023/11.txt").iter().map(|l| l.chars().collect_vec()).collect_vec();
     let (w, h) = (l[0].len(), l.len());
     let s = (0..w).cartesian_product(0..h).filter(|&(x, y)| l[y][x] == '#').collect_vec();
@@ -13,7 +13,7 @@ fn solve(factor: i64) -> usize {
             let (y1, y2) = if s[a].1 < s[b].1 { (s[a].1, s[b].1) } else { (s[b].1, s[a].1) };
             let lx = ex.partition_point(|x| s[b].0 > *x) - ex.partition_point(|x| s[a].0 > *x);
             let ly = ey.partition_point(|y| y2 > *y) - ey.partition_point(|y| y1 > *y);
-            (s[b].0 - s[a].0) + (y2 - y1) + (lx + ly) * (factor as usize - 1)
+            (s[b].0 - s[a].0) + (y2 - y1) + (lx + ly) * (factor - 1)
         })
         .sum::<usize>()
 }
